Primary keys changed from Integer to BigInteger

+++ haas/ext/switches/mock.py
@@ -21,7 +21,7 @@ from collections import defaultdict
 from haas.model import Switch
 from haas.migrations import paths
 import schema
-from sqlalchemy import Column, Integer, ForeignKey, String
+from sqlalchemy import Column, BigInteger, ForeignKey, String
 from os.path import dirname, join
 
 paths[__name__] = join(dirname(__file__), 'migrations', 'mock')
@@ -43,7 +43,7 @@ class MockSwitch(Switch):
         'polymorphic_identity': api_name,
     }
 
-    id = Column(Integer, ForeignKey('switch.id'), primary_key=True)
+    id = Column(BigInteger, ForeignKey('switch.id'), primary_key=True)
     hostname = Column(String, nullable=False)
     username = Column(String, nullable=False)
     password = Column(String, nullable=False)
